+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+7.10.4
+======
+
+*   (improvement) Add helper `DeferredRoute::ensureValidValue()`.
+*   (improvement) Add helper `DeferredTranslation::ensureValidValue()`.
+
+
 7.10.3
 ======